Changed ColorComponent members type from 4 bytes(float) to 1 byte(unsigned char).

+++ include/ECS/Component/ColorComponent.h
@@ -1,17 +1,19 @@
 #include <string>
+#include "ECS/Component/Component.h"
+#include "common/types.h"
 
 namespace ECS::Component
 {
     class ColorComponent : public Component
     {
     public:
-            ColorComponent(int r, int g, int b, int a) : Component("Color"), r(r), g(g), b(b), a(a) {}
+            ColorComponent(u8 r, u8 g, u8 b, u8 a) : Component("Color"), r(r), g(g), b(b), a(a) {}
             ~ColorComponent() {}
 
-            int r;
-            int g;
-            int b;
-            int a;
+            u8 r;
+            u8 g;
+            u8 b;
+            u8 a;
 
     };
 }
